fix: fixed button variants issue
